# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: NOMOS API
  version: 0.1.0
  description: |
    에이전트 협업 조율 서버. 수동 테스트 절차는 `docs/manual-test.md`를 함께 보라.

    **토큰이 두 종류다.** 오른쪽 위 Authorize를 누르면 칸이 두 개 뜬다.

    | 칸 | 어떤 토큰 | 받는 법 |
    |---|---|---|
    | `userToken` | 사람 토큰 (`kind: user`) | `POST /auth/login` — 시드 계정 `rep` / `manual-test-1234` |
    | `agentToken` | 에이전트 토큰 (`kind: agent`) | `npm run seed` 출력의 **be-laptop access token**, 또는 `POST /agents/token/refresh` |

    서로의 API를 부를 수 없다. 사람 토큰으로 태스크 API를 부르면 401이다.

    access token은 **1시간**이다. 만료되면 `POST /agents/token/refresh`로 다시 받는다.
servers:
  - url: http://localhost:3000/api
    description: 로컬 개발 서버
components:
  securitySchemes:
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 사람 토큰. `POST /auth/login`의 `data.accessToken`을 붙여넣는다.
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 에이전트 토큰. 시드 출력의 access token 또는 refresh 응답의 값을 붙여넣는다.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: FORBIDDEN_PATH
            message:
              type: string
              example: '**/.env* is denied'
            reason:
              type: string
              description: POLICY_STALE에만 실린다. 다른 코드에서는 생략된다.
              example: policy_stale
  responses:
...
    PolicyStale:
      description: |
        토큰이 **옛 정책 스냅샷 기준**이다. 경로 규칙이나 정책이 바뀌면 이렇게 된다.
        refresh로 재발급하고 원 요청을 **1회만** 재시도한다. 그래도 stale이면 멈춘다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: POLICY_STALE
              message: policy snapshot changed; refresh the token
              reason: policy_stale
...
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{TASK_BE}}'
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{PROJECT_ID}}'
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: 산출물 id. `GET /tasks/{taskId}/artifacts`로 찾는다.
...
paths:
...
  /tasks/{taskId}/claim:
    post:
      tags: [태스크·노트]
      summary: claim_task — READY 태스크 선점
      description: |
        `state='READY' AND assignee_agent_id IS NULL` 조건부 UPDATE 하나로 경합을 가른다.
        **낙관적 잠금이지 트랜잭션 잠금이 아니다.** 진 쪽은 0행을 받고 그게 곧 409다.

        먼저 확인하는 것: 프로젝트 멤버인가 → 선행 태스크가 전부 DONE인가 → 내 역할과 맞는가.
        `tasks.team_role`이 NULL이면 역할 제한이 없다(통합 태스크).
      security: [{ agentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '200':
          description: 선점됨
          content:
            application/json:
              example:
                data: { id: '{{TASK_BE}}', state: CLAIMED, assigneeAgentId: 2012e1df-0beb-43fd-ac09-d893ba56b01e, teamRole: BACKEND }
        '401':
          $ref: '#/components/responses/PolicyStale'
        '403':
          description: 역할 불일치·프로젝트 멤버 아님·프로젝트 정지
          content:
            application/json:
              examples:
                역할불일치:
                  value: { error: { code: TASK_ROLE_MISMATCH, message: task belongs to FRONTEND } }
                정지됨:
                  value: { error: { code: PROJECT_HALTED, message: project is halted } }
        '404':
          description: 없는 태스크 (다른 프로젝트의 태스크도 "없음"으로 답한다)
          content:
            application/json:
              example:
                error: { code: TASK_NOT_FOUND, message: task not found }
        '409':
          description: 이미 선점됨 또는 선행 의존 미완료
          content:
            application/json:
              examples:
                경합패배:
                  value: { error: { code: TASK_ALREADY_CLAIMED, message: task is no longer READY } }
                의존미완료:
                  value: { error: { code: TASK_DEPS_NOT_DONE, message: 1 dependencies are not DONE } }
  /tasks/{taskId}/artifacts:
    post:
      tags: [태스크·노트]
      summary: submit_artifact — 산출물 제출
      description: |
        제출 시점에 경로를 **다시** 판정한다. 로컬 편집기로 직접 고친 파일은 도구 호출을 거치지 않으므로
        여기서만 걸린다(V3의 전신).

        걸린 행동 키(`triggeredActions`)와 **가장 엄격한 판정**(`gateMode`)을 `artifacts`에 고정한다 —
        나중에 정책표나 경로 규칙이 바뀌어도 "그때 무엇에 걸렸나"가 남아야 리플레이가 성립한다.

        `changedPaths`에 `migrations/010_x.sql`을 섞으면 `db:migration`이 걸려 L2에서 `gateMode=HUMAN`이 된다.
      security: [{ agentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [commitSha, changedPaths]
              properties:
                commitSha: { type: string, pattern: '^[0-9a-f]{7,40}$' }
                changedPaths: { type: array, minItems: 1, items: { type: string } }
            examples:
              정상:
                value: { commitSha: a1b2c3d, changedPaths: [src/participation.ts, tests/participation.test.ts] }
              금지경로:
                value: { commitSha: a1b2c3d, changedPaths: [api/.env] }
              승인필요:
                value: { commitSha: a1b2c3d, changedPaths: [migrations/010_x.sql] }
      responses:
        '201':
          description: |
            제출됨. 태스크는 VERIFYING으로 넘어간다.

            **서버가 판정하는 단계(V1A·V1B·V3)는 이 응답 안에서 이미 끝나 있다.** 비동기로 돌리면
            "제출은 됐는데 결과는 언제 오는가"를 관리하는 상태가 하나 더 늘고, 모델은 그 사이에
            다음 태스크로 넘어가 버린다.

            `verification.outcome`이 읽는 법이다 —
            `PENDING`(V2·V4 보고를 기다린다) · `DONE` · `AWAITING_APPROVAL`(gateMode가 AUTO가 아니다) ·
            `RETRY`(FAIL — 태스크가 READY로 돌아가고 retry_count가 올랐다) · `ESCALATED`(3회째 FAIL).
          content:
            application/json:
              example:
                data:
                  id: 44444444-4444-4444-4444-444444444444
                  taskId: '{{TASK_BE}}'
                  commitSha: a1b2c3d
                  triggeredActions: ['artifact:submit', 'code:own_path', 'test:write']
                  gateMode: AUTO
                  attempt: 1
                  verification:
                    artifactId: 44444444-4444-4444-4444-444444444444
                    attempt: 1
                    stages:
                      - { stage: V1A, result: SKIPPED }
                      - { stage: V1B, result: SKIPPED }
                      - { stage: V3, result: PASS }
                    outcome: PENDING
                    taskState: VERIFYING
                    retryCount: 0
        '401':
          $ref: '#/components/responses/PolicyStale'
        '403':
          description: |
            금지 경로·소유권 위반·남의 태스크. 세 경우 모두 `artifacts`에는 행이 생기지 않고
            `TOOL_DENIED` 이벤트만 남는다. 소유권 위반만 `path_violation=true`다.
          content:
            application/json:
              examples:
                금지경로:
                  value: { error: { code: FORBIDDEN_PATH, message: '**/.env* is denied' } }
                소유권위반:
                  value: { error: { code: SCOPE_DENIED, message: '** belongs to FRONTEND' } }
                남의태스크:
                  value: { error: { code: NOT_TASK_ASSIGNEE, message: task is assigned to another agent } }
        '409':
          description: CLAIMED·IN_PROGRESS가 아닌 상태
          content:
            application/json:
              example:
                error: { code: TASK_STATE_INVALID, message: cannot submit while task is VERIFYING }

    get:
      tags: [검증]
      summary: 산출물 목록 (최신 먼저)
      description: |
        모델은 MCP 도구로 제출하므로 Executor는 방금 생긴 산출물의 id를 모른다. V2·V4 결과를
        어디에 붙일지 알아야 해서 목록을 준다.

        **사람도 호출한다** — 웹 UI가 태스크 상세에서 제출 이력을 보고, 여기서 얻은 id로
        `GET /artifacts/{artifactId}/verifications`에 닿는다. 사람 경로가 없으면 그 화면이 성립하지 않는다.

        접근 범위가 다르다: 에이전트는 **담당자만**, 사람은 대표이거나 그 프로젝트에 배정된
        에이전트의 주인이면 된다(조회뿐이다).
      security: [{ agentToken: [] }, { userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '200':
          description: attempt 내림차순
          content:
            application/json:
              example:
                data:
                  artifacts:
                    - { id: 44444444-4444-4444-4444-444444444444, taskId: '{{TASK_BE}}', commitSha: a1b2c3d, attempt: 2, gateMode: AUTO }
...
  /projects/{projectId}/tasks:
    get:
      tags: [태스크·노트]
      summary: 태스크 목록
      description: |
        웹 UI(사람 토큰)와 Executor 폴링(에이전트 토큰)이 **같은 엔드포인트**를 쓴다.

        에이전트 토큰이면 `teamRole` 질의를 무시하고 자기 역할로 강제 필터한다 — 남의 역할 태스크가
        목록에 아예 없어야 "잡아보고 403"을 반복하지 않는다.
      security: [{ agentToken: [] }, { userToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - { name: state, in: query, schema: { type: string, enum: [READY, CLAIMED, IN_PROGRESS, VERIFYING, AWAITING_APPROVAL, BLOCKED, ESCALATED, DONE] } }
        - { name: teamRole, in: query, schema: { type: string, enum: [FRONTEND, BACKEND] }, description: 에이전트 토큰이면 무시된다 }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: 생성 순서대로. 정렬이 결정적이어야 폴링이 같은 태스크를 두고 경합하지 않는다.
          content:
            application/json:
              example:
                data:
                  tasks:
                    - { id: '{{TASK_BE}}', title: 'T-001 참여 신청 API 구현', state: READY, teamRole: BACKEND, retryCount: 0 }
  /tasks/{taskId}/briefing:
    get:
      tags: [태스크·노트]
      summary: 태스크 브리핑 — 프롬프트·settings.json 재료 한 번에
      description: |
        Executor는 팀원 노트북에서 돌아 DB를 직접 못 본다. 여러 엔드포인트를 조합하게 하면 조합 방식이
        Executor마다 갈리므로 서버가 한 번에 준다 — 명세, 서버가 렌더한 인계 노트 블록(`notesBlock`),
        쓸 수 있는 경로, `.claude/settings.json`에 그대로 깔 `claudeSettings`, 그리고 V2가 돌릴 `specTests`.

        **CLAIM 전에도 받을 수 있다.** Executor가 작업공간을 먼저 만들고, claim은 모델이 도구로 하기 때문이다.
        남의 역할 태스크는 403이다 — 명세·노트를 미리 보지 못하게 한다.
      security: [{ agentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                data:
                  task: { id: '{{TASK_BE}}', title: 'T-001 참여 신청 API 구현', state: READY }
                  repo: { id: '{{REPO_API}}', fullName: acme/study-api, defaultBranch: main, devBranch: dev }
                  spec: { featureKey: F-03, title: 참여 신청 API, content: 'WHEN 정원이 찬 스터디에 신청하면 THEN 409를 반환한다' }
                  notesBlock: "[인계 노트]\n#1 [명세 이탈] quota check moved to service\n  · handler was too late\n  영향: contracts/F-03.yaml"
                  writablePaths: [{ pathPattern: 'src/**', access: write, ownerRole: BACKEND }]
                  claudeSettings:
                    permissions:
                      deny: ['Read(**/.env*)', 'Read(!**/.env.example)', 'Edit(/contracts/**)', 'Edit(**/secrets/**)']
                  specTests: [{ id: 5f0e..., criterion: 정원이 찬 스터디 신청은 409다, lockedAt: '2026-09-21T00:00:00.000Z' }]
                  policyHash: '{{POLICY_HASH}}'
  /tasks/{taskId}/branch:
    patch:
      tags: [태스크·노트]
      summary: 작업 브랜치 보고
      description: |
        Executor가 worktree를 만들면서 정한 브랜치 이름을 서버에 알린다. 서버가 정본을 들고 있어야
        재실행·이어받기에서 같은 브랜치를 쓰고 나중에 PR을 열 때도 이 값을 본다.

        **이미 값이 있으면 덮어쓰지 않는다** — 브랜치가 바뀌면 이전 작업물을 찾을 수 없게 된다.
      security: [{ agentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branchName]
              properties:
                branchName: { type: string, maxLength: 200, pattern: '^[A-Za-z0-9._/-]+$' }
            example: { branchName: 'task/1d5c8f2a' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                data: { id: '{{TASK_BE}}', branchName: 'task/1d5c8f2a', state: CLAIMED }
  /artifacts/{artifactId}/verifications:
    post:
      tags: [검증]
      summary: 브릿지 단계(V2·V4) 보고
      description: |
        V2(PM 시험지)와 V4(린트)는 작업공간이 있어야 돌 수 있어 제출 시점에는 결과가 물리적으로
        존재할 수 없다. 그래서 이 단계만 브릿지가 나중에 보고한다.

        **V1A·V1B·V3은 받지 않는다**(403 `VERIFICATION_STAGE_NOT_REPORTABLE`). 로컬이 보고한 V3를
        서버가 받으면 "로컬을 신뢰하지 않는다"는 V3의 전제가 무너진다.

        `SKIPPED`는 반드시 `detail.reason`을 함께 보낸다 — 사유 없는 SKIPPED는 나중에 PASS와 구분되지 않는다.

        V2를 `PASS`로 보고해도 서버가 `spec_tests.locked_at < artifacts.created_at`을 다시 확인하고,
        산출물보다 늦게 잠긴 시험지가 있으면 `FAIL`로 뒤집는다.
      security: [{ agentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/ArtifactId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage, result]
              properties:
                stage: { type: string, enum: [V2, V4] }
                result: { type: string, enum: [PASS, FAIL, SKIPPED] }
                detail: { type: object }
                durationMs: { type: integer, minimum: 0 }
            examples:
              린트통과:
                value: { stage: V4, result: PASS, detail: { command: 'npm run lint' }, durationMs: 4200 }
              시험지없음:
                value: { stage: V2, result: SKIPPED, detail: { reason: '작업공간에 vitest가 설치돼 있지 않다' } }
              린트실패:
                value: { stage: V4, result: FAIL, detail: { exitCode: 1, output: "src/index.ts:3:1 error" } }
      responses:
        '201':
          description: |
            기록됨. 응답은 이 산출물의 **전체** 결론이다 — 필요한 단계(V3·V2·V4)가 모두 들어오고
            하나도 FAIL이 아니면 `DONE`(gateMode가 AUTO가 아니면 `AWAITING_APPROVAL`)이 된다.
          content:
            application/json:
              example:
                data:
                  artifactId: 44444444-4444-4444-4444-444444444444
                  attempt: 1
                  stages:
                    - { stage: V1A, result: SKIPPED }
                    - { stage: V1B, result: SKIPPED }
                    - { stage: V2, result: SKIPPED }
                    - { stage: V3, result: PASS }
                    - { stage: V4, result: PASS }
                  outcome: DONE
                  taskState: DONE
                  retryCount: 0
        '400':
          description: 사유 없는 SKIPPED
          content:
            application/json:
              example:
                error: { code: VALIDATION_ERROR, message: SKIPPED must carry detail.reason }
        '403':
          description: 서버가 판정하는 단계이거나, 남의 태스크다
          content:
            application/json:
              examples:
                서버판정단계:
                  value: { error: { code: VERIFICATION_STAGE_NOT_REPORTABLE, message: stage V3 is judged by the server } }
                남의태스크:
                  value: { error: { code: NOT_TASK_ASSIGNEE, message: task is assigned to another agent } }
        '409':
          description: |
            같은 단계를 두 번 보고했다. 재제출은 attempt가 올라간 **새 산출물**이므로, FAIL을 PASS로
            갈아치우는 경로는 없다.
          content:
            application/json:
              example:
                error: { code: VERIFICATION_ALREADY_RECORDED, message: V4 is already recorded for this artifact }
    get:
      tags: [검증]
      summary: 검증 결과 조회
      description: 무엇이 PASS였고 무엇을 **왜** 못 돌렸는지 본다. SKIPPED의 `detail.reason`이 그 답이다.
      security: [{ agentToken: [] }, { userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/ArtifactId' }]
      responses:
        '200':
          description: stage 순
          content:
            application/json:
              example:
                data:
                  verifications:
                    - stage: V1A
                      result: SKIPPED
                      executedBy: server
                      detail: { reason: 'contracts 테이블 미구현 — ⑦ 이후 가능' }
                    - stage: V3
                      result: FAIL
                      executedBy: server
                      durationMs: 732
                      detail:
                        reason: 'changed_paths가 커밋의 실제 diff와 다르다'
                        undeclared: [README.md]
                        overdeclared: []